requirements.txt
# OpenAI
openai==0.27.0

# Basic requirements
python-dotenv==0.20.0

# Streamlit
streamlit==1.10.0